--- src/attention/attention.py
@@ -1,1 +1,0 @@
-import torch
